Rearrange middleware to allow users to be sanitized of passwords in all routes
